bpftop,metrics: separate workspace for bpftop app & ebpf-metrics lib
This separates the project workspace into the bpftop app/bin and
ebpf-metrics lib. The goal of ebpf-metrics is to be a library that
provides an "protocol-agnostic" interface for ebpf metrics emission.
This is done with the `prometheus-client` crate, which despite the name,
strictly implements the OpenMetrics standard for sending and receiving
metrics.

This PR also aims to migrate all libbpf deps to aya, as aya supports
kernel-space programs written in Rust. This could remove the need of a
custom build script for translating the C bpf program to Rust.